+++ .eslintrc.cjs
@@ -11,6 +11,8 @@ module.exports = {
       files: ['**/*.ts'],
       excludedFiles: '**/__tests__/**/*.ts',
       rules: {
+        // Disallow importing a node module without it being specified in package.json
+        'import/no-extraneous-dependencies': 'error',
         'import/extensions': ['error', 'ignorePackages'],
         '@typescript-eslint/consistent-type-imports': [
           'error',